requirements.txt
acpype @ file:///home/conda/feedstock_root/build_artifacts/acpype_1623704583461/work
AmberLite==16.0
AmberTools==20.15
ase==3.22.0
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1625835293160/work
cftime @ file:///home/conda/feedstock_root/build_artifacts/cftime_1621553714314/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1610093490430/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1621503718520/work
combo==0.2
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1616851476134/work
cycler==0.10.0
Cython @ file:///tmp/build/80754af9/cython_1618435160151/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1621187651333/work
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1621629915293/work
future @ file:///home/conda/feedstock_root/build_artifacts/future_1610147327521/work
gmpy2==2.1.0b5
greenlet @ file:///home/conda/feedstock_root/build_artifacts/greenlet_1620327505524/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
importlib-metadata==4.6.3
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1621360792037/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1621419064915/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1610099769230/work
latexcodec @ file:///home/conda/feedstock_root/build_artifacts/latexcodec_1592937263153/work
llvmlite==0.36.0
loguru @ file:///home/conda/feedstock_root/build_artifacts/loguru_1618589835436/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1621455677251/work
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1620506037462/work
MMPBSA.py==16.0
monty @ file:///home/conda/feedstock_root/build_artifacts/monty_1625794779766/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1612895720168/work
netCDF4 @ file:///home/conda/feedstock_root/build_artifacts/netcdf4_1618854182283/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1598210780226/work
numba==0.53.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1624399863733/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packmol-memgen==1.1.0rc0
palettable==3.3.0
pandas==1.3.1
ParmEd @ file:///home/conda/feedstock_root/build_artifacts/parmed_1618283661161/work
pdb4amber==1.7.dev0
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1621287973719/work
plotly @ file:///home/conda/feedstock_root/build_artifacts/plotly_1624919128282/work
py3Dmol==0.9.1
pyberny==0.6.3
pybtex @ file:///home/conda/feedstock_root/build_artifacts/pybtex_1611054464752/work
pycairo==1.20.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pymatgen @ file:///home/conda/feedstock_root/build_artifacts/pymatgen_1625177766451/work
pyMSMT==20.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1610291447907/work
python-dateutil==2.8.1
pytraj==2.0.5
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
pyxtal @ git+https://github.com/qzhu2017/PyXtal.git@2dd0fbf5173ed66043430421a2d31639f1de4a93
PyYAML==5.4.1
reportlab==3.5.68
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
ruamel.yaml @ file:///home/conda/feedstock_root/build_artifacts/ruamel.yaml_1624567361774/work
ruamel.yaml.clib @ file:///home/conda/feedstock_root/build_artifacts/ruamel.yaml.clib_1610146840614/work
sander==16.0
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1624824888010/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
spglib @ file:///home/conda/feedstock_root/build_artifacts/spglib_1611910190481/work/python
SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/sqlalchemy_1624920405831/work
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1618015367433/work
tabulate @ file:///home/conda/feedstock_root/build_artifacts/tabulate_1614001031686/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1625742961149/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094706440/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1625664838498/work
uncertainties @ file:///home/conda/feedstock_root/build_artifacts/uncertainties_1605702954409/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1624634538755/work
vtk==9.0.0
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1621518206714/work
xtb-python @ file:///home/conda/feedstock_root/build_artifacts/xtb-python_1620120837566/work
zipp==3.5.0
